backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis for caching
redis==5.0.1
hiredis==2.3.2

# Celery for async tasks
celery==5.3.4
flower==2.0.1

# AI/ML
anthropic==0.8.1
openai==1.8.0
langchain==0.1.0
pandas==2.1.4
numpy==1.26.3

# Payment processing
stripe==7.9.0

# Security
cryptography==42.0.0
itsdangerous==2.1.2

# Monitoring and logging
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
sentry-sdk[fastapi]==1.39.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.2.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# File handling
python-magic==0.4.27
aiofiles==23.2.1

# Email
aiosmtplib==3.0.1
email-validator==2.1.0

# Utils
python-dateutil==2.8.2
pytz==2023.3
tenacity==8.2.3